fix(mac): extract IndieArtist model to shared file so macOS build can resolve the type
IndieArtist was defined inside IndieArtistDirectoryService.swift which
has platformFilters = (ios,) in the Xcode project. UnstreamAPI.swift
(no platform filter, compiled for all targets) referenced IndieArtist,
causing macOS builds to fail with "Cannot find type 'IndieArtist' in
scope" and a cascading Decodable conformance error on the local
DirectoryResponse struct.

Fix: move IndieArtist into a new Models/IndieArtist.swift with no
platform filter so it is visible to both macOS and iOS compilation
units. Remove the duplicate definition from IndieArtistDirectoryService.
