+++ apps/mac/Unstream/Models/IndieArtist.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+struct IndieArtist: Codable, Identifiable, Equatable {
+    let slug: String
+    let name: String
+    let imageUrl: String?
+
+    var id: String { slug }
+}
